分析器代码整理
